requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.86.0
anyio==4.12.1
async-timeout==5.0.1
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
docstring_parser==0.17.0
duckduckgo_search==8.1.1
exceptiongroup==1.3.1
fastapi==0.135.1
fonttools==4.62.1
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
langchain==1.2.13
langchain-anthropic==1.4.0
langchain-core==1.2.20
langchain-openai==1.1.11
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.8
multidict==6.7.1
narwhals==2.18.0
numpy==2.2.6
openai==2.29.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==2.3.3
pillow==12.1.1
plotly==6.6.0
primp==1.1.3
propcache==0.4.1
protobuf==6.33.6
pyarrow==23.0.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
pyparsing==3.3.2
pypdf==6.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==4.0.0
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
reportlab==4.4.10
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
starlette==0.52.1
streamlit==1.55.0
tavily-python==0.7.23
tenacity==9.1.4
tiktoken==0.12.0
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.42.0
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
